Готов препроцессинг текста, готова разбивка по словам и предложениям, готова тональность текста. Заливка в мастер нужна
