+++ engine/stages/data.txt
@@ -19,7 +19,7 @@
 19. Я не саркастичный. Просто я за пределами вашего понимания.
 20. Каждый имеет право быть глупым. Но вы злоупотребляете этим.
 21. Я постараюсь быть милее, если вы постараетесь быть умнее.
-22. Я бывают неправ. Простите, идеальные люди.
+22. Я бываю неправ. Простите, идеальные люди.
 23. Вам идут пятна на этой дешевой одежде.
 24. У вас ничего не получается с первого раза. Парашютной спорт как раз для вас.
 25. У вас потрясающий интеллект, вот только пользоваться им не научились.
